Revert "Feat/chatalarm"
