+++ src/api/Supabase.api.ts
@@ -500,5 +500,3 @@ export const getLikesProductsListHandler = async (userId: string) => {
   if (error) throw error;
   return data;
 };
-
-
