--- src/components/chat/market/TotalChatUnreadCounts.type.ts
@@ -1,4 +1,0 @@
-export type TotalChatUnreadCountsProps = {
-    ChatBtnOpen: boolean;
-  };
-
