+++ src/pages/MarketProduct.tsx
@@ -1,7 +1,6 @@
 import Product from '../components/market/marketproduct/Product';
 
 const MarketProduct = () => {
-  console.log('상세화면')
   return <Product />;
 };
