+++ src/state/atom/chatAtom.ts
@@ -1,5 +1,7 @@
 import { atom } from 'recoil';
-import { ChatRoomTypes, UnreadCount } from './Chat.type';
+import { UnreadCount } from '../../App';
+import { MessageTypes } from '../../types/types';
+import { ChatRoomTypes } from './Chat.type';
 
 export const productState = atom({
   key: 'productState', // 고유한 키
@@ -11,58 +13,17 @@ export const person = atom({
   default: '', // 기본값
 });
 
+export const otherPerson = atom({
+  key: 'otherPerson', // 고유한 키
+  default: '', // 기본값
+});
+
 export const ChatId = atom({
   key: 'ChatId', // 고유한 키
   default: '', // 기본값
 });
 
-
-export const mainChatModalOpen = atom({
-  key: 'mainChatModalOpen', // 고유한 키
-  default: false, // 기본값
-});
-export type productDetails = {
-  image: string;
-  title: string;
-  price: number;
-  id: number;
-};
-export type otherUserDetails = {
-  nickname: string;
-  user_img: string;
-};
-
-
-export const otherUserDetail = atom<otherUserDetails | null>({
-  key: 'otherUserDetail', 
-  default: null, // Now null is a valid default value
-});
-
-export const productDetail = atom<productDetails | null>({
-  key: 'productDetail',
-  default: null, // Now null is a valid default value
-});
-
-export type MessageType = {
-  id: number;
-  content: string;
-  author_id: string;
-  chat_id: string;
-  item_id: number;
-  others_id: string;
-  users?: UserType; // 사용자 닉네임을 포함할 수 있는 옵셔널 프로퍼티
-  created_at: number;
-};
-
-export type UserType = {
-  id: string;
-  email: string;
-  lastMessage?: string; // lastMessage 속성 추가 (옵셔널로 처리)
-  nickname: string;
-};
-
-
-export const sendMessages = atom<MessageType[]>({
+export const sendMessages = atom<MessageTypes[]>({
   key: 'sendMessagesState',
   default: [], // 초기값은 빈 배열
 });
@@ -72,10 +33,8 @@ export const chatRoomsState = atom<ChatRoomTypes[]>({
   default: [], // 초기값은 빈 배열
 });
 
-
-
-export const firstChatModalOpenState = atom({
-  key: 'firstChatModalOpenState',
+export const isChatModalOpenState = atom({
+  key: 'isChatModalOpenState',
   default: false,
 });
 
@@ -99,75 +58,7 @@ export const loginUserState = atom({
   default: '',
 });
 
-export const selectedUserImages = atom({
-  key: 'selectedUserImages',
-  default: '',
-});
-
 export const UnreadCounts = atom<UnreadCount[]>({
   key: 'UnreadCounts',
   default: [],
 });
-
-
-export const updateMesaages = atom<MessagePayload>({
-  key: 'updateMesaages',
-  default:{
-    commit_timestamp: '',
-    errors: null , // Replace 'any' with a more specific type if possible
-    eventType: '',
-    new: {
-      author_id: '',
-      chat_id: '',
-      content: '',
-      created_at: '',
-      id: 0,
-      item_id: 0,
-      others_id: '',
-    },
-    
-    old: '', // Replace 'any' with a more specific type if possible
-    schema: '',
-    table: '',
-  }
-   
-});
-export type MessagePayload = {
-  commit_timestamp: string;
-  errors: null | any; // Replace 'any' with a more specific type if possible
-  eventType: string;
-  new: {
-    author_id: string;
-    chat_id: string;
-    content: string;
-    created_at: string;
-    id: number;
-    item_id: number;
-    others_id: string | null;
-  };
-  
-  old?: any; // Replace 'any' with a more specific type if possible
-  schema: string;
-  table: string;
-};
-
-
-
-// // ======================================
-// // 사용법
-// export const textState = atom({
-//   key: 'textState', // 고유한 키
-//   default: '', // 기본값
-// });
-
-// export const charCountState = selector({
-//   key: 'charCountState', // 고유한 키
-//   get: ({get}) => {
-//     const text = get(textState);
-//     return text.length;
-//   },
-// });
-
-// 쓰는 법
-// const [productId, setProductId] = useRecoilState(productState);
-// const productId = useRecoilValue(productState);
